Refactor underway
