+++ rlm_kanidm/module/include/freeradius_wrapper.h
@@ -35,18 +35,5 @@ typedef struct rlm_kanidm_conf_parser_t {
     void const *dflt;
 } rlm_kanidm_conf_parser_t;
 
-typedef int (*rlm_kanidm_instantiate_t)(void *mod_cs, void *instance);
-typedef int (*rlm_kanidm_detach_t)(void *instance);
-typedef rlm_rcode_t (*rlm_kanidm_packetmethod_t)(void *instance, REQUEST *request);
 
-typedef struct rlm_kanidm_module_t {
-    uint64_t magic;
-    char const *name;
-    int type;
-    size_t inst_size;
-    rlm_kanidm_conf_parser_t const *config;
-    rlm_kanidm_instantiate_t bootstrap;
-    rlm_kanidm_instantiate_t instantiate;
-    rlm_kanidm_detach_t detach;
-    rlm_kanidm_packetmethod_t methods[RLM_KANIDM_MOD_COUNT];
-} rlm_kanidm_module_t;
+
