+++ rlm_kanidm/module/src/lib.rs
@@ -36,19 +36,12 @@ use tracing::error;
 
 #[cfg(feature = "extern-freeradius-module")]
 mod freeradius;
+#[cfg(feature = "extern-freeradius-module")]
+pub mod syms;
 
+mod cache;
+mod error;
 mod ffi;
-
-const ATTR_USER_NAME: &str = "User-Name";
-const ATTR_TLS_CN: &str = "TLS-Client-Cert-Common-Name";
-const ATTR_TLS_SAN_DN_CN: &str = "TLS-Client-Cert-Subject-Alt-Name-Directory-Name-Common-Name";
-
-const REPLY_USER_NAME: &str = "User-Name";
-const REPLY_MESSAGE: &str = "Reply-Message";
-const REPLY_TUNNEL_TYPE: &str = "Tunnel-Type";
-const REPLY_TUNNEL_MEDIUM_TYPE: &str = "Tunnel-Medium-Type";
-const REPLY_TUNNEL_PRIVATE_GROUP_ID: &str = "Tunnel-Private-Group-ID";
-const CONTROL_CLEARTEXT_PASSWORD: &str = "Cleartext-Password";
 
 /// RADIUS response codes as expected by FreeRADIUS.
 /// ```rust
@@ -147,124 +140,9 @@ impl RequestAttributes {
     }
 }
 
-#[derive(Debug)]
-struct CacheEntry {
-    token: RadiusAuthToken,
-    fetched_at: Instant,
-}
 
-impl CacheEntry {
-    /// Duration since this entry was pulled from Kanidm
-    fn age(&self, now: Instant) -> Duration {
-        now.saturating_duration_since(self.fetched_at)
-    }
-
-    /// If it's less than the TTL
-    fn fresh(&self, now: Instant, ttl: Duration) -> bool {
-        self.age(now) <= ttl
-    }
-
-    /// If it's past the TTL but within the stale window
-    fn stale_allowed(&self, now: Instant, ttl: Duration, stale_window: Duration) -> bool {
-        let age = self.age(now);
-        age > ttl && age <= ttl.saturating_add(stale_window)
-    }
-}
-
-#[derive(Debug)]
-pub enum ModuleError {
-    Io(String),
-    Config(String),
-    Http(String),
-    Other(String),
-}
-
-impl std::fmt::Display for ModuleError {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Self::Io(s) => write!(f, "IO Error: {s}"),
-            Self::Config(s) => write!(f, "Config Error: {s}"),
-            Self::Http(s) => write!(f, "HTTP Error: {s}"),
-            Self::Other(s) => write!(f, "Internal Error: {s}"),
-        }
-    }
-}
-
-impl From<ModuleError> for AuthResultC {
-    fn from(input: ModuleError) -> AuthResultC {
-        auth_error(&Response::Fail, input.to_string())
-    }
-}
-
-impl std::error::Error for ModuleError {}
-
-struct LookupCache {
-    cache_ttl: Duration,
-    cache_stale_if_error: Duration,
-    cache_max_entries: usize,
-    cache: Mutex<BTreeMap<String, CacheEntry>>,
-}
-
-impl LookupCache {
-    fn new() -> Self {
-        Self {
-            cache_ttl: Duration::from_secs(30),
-            cache_stale_if_error: Duration::from_secs(120),
-            cache_max_entries: 10_000,
-            cache: Mutex::new(BTreeMap::new()),
-        }
-    }
-
-    fn lookup_cache(&self, user_id: &str, now: Instant) -> Option<RadiusAuthToken> {
-        let Ok(mut cache_guard) = self.cache.lock() else {
-            error!("Couldn't acquire cache lock for lookup_cache");
-            return None;
-        };
-        let entry = cache_guard.get(user_id)?;
-        if entry.fresh(now, self.cache_ttl) {
-            return Some(entry.token.clone());
-        } else {
-            cache_guard.remove(user_id);
-        }
-        None
-    }
-
-    fn lookup_stale_cache(&self, user_id: &str, now: Instant) -> Option<RadiusAuthToken> {
-        let Ok(mut cache_guard) = self.cache.lock() else {
-            error!("Couldn't acquire cache lock for lookup_stale_cache");
-            return None;
-        };
-        let entry = cache_guard.get(user_id)?;
-        if entry.stale_allowed(now, self.cache_ttl, self.cache_stale_if_error) {
-            return Some(entry.token.clone());
-        } else {
-            cache_guard.remove(user_id);
-        }
-        None
-    }
-
-    fn insert_cache(&self, user_id: String, token: RadiusAuthToken, now: Instant) {
-        if let Ok(mut guard) = self.cache.lock() {
-            if guard.len() >= self.cache_max_entries && !guard.contains_key(&user_id) {
-                if let Some(oldest_key) = guard
-                    .iter()
-                    .min_by_key(|(_, entry)| entry.fetched_at)
-                    .map(|(k, _)| k.clone())
-                {
-                    guard.remove(&oldest_key);
-                }
-            }
-            guard.insert(
-                user_id,
-                CacheEntry {
-                    token,
-                    fetched_at: now,
-                },
-            );
-        } else {
-            error!("Couldn't acquire cache lock for insert_cache");
-        }
-    }
+pub struct ModuleHandle {
+    module: Arc<Module>,
 }
 
 pub struct Module {
@@ -422,198 +300,6 @@ impl Module {
             }
         }
     }
-}
-
-pub struct ModuleHandle {
-    module: Arc<Module>,
-}
-
-#[repr(C)]
-pub struct KVPair {
-    pub key: *const c_char,
-    pub value: *const c_char,
-}
-
-#[repr(C)]
-pub struct AuthResultC {
-    pub code: i32,
-    pub reply: *mut KVPair,
-    pub reply_len: usize,
-    pub control: *mut KVPair,
-    pub control_len: usize,
-    pub error: *mut c_char,
-}
-
-pub(crate) fn cstr_to_string(ptr_in: *const c_char) -> Result<String, ModuleError> {
-    if ptr_in.is_null() {
-        return Err(ModuleError::Other("null string pointer".to_string()));
-    }
-    let cstr = unsafe { CStr::from_ptr(ptr_in) };
-    cstr.to_str()
-        .map(|s| s.to_string())
-        .map_err(|e| ModuleError::Other(format!("invalid utf-8 string: {e}")))
-}
-
-fn auth_result_from_pairs(response: &Response) -> AuthResultC {
-    let reply_vec = into_kvpairs(response.reply());
-    let control_vec = into_kvpairs(response.control());
-    let reply_len = reply_vec.len();
-    let control_len = control_vec.len();
-    let mut reply_boxed = reply_vec.into_boxed_slice();
-    let mut control_boxed = control_vec.into_boxed_slice();
-    let reply_ptr = reply_boxed.as_mut_ptr();
-    let control_ptr = control_boxed.as_mut_ptr();
-    std::mem::forget(reply_boxed);
-    std::mem::forget(control_boxed);
-    AuthResultC {
-        code: response.code(),
-        reply: reply_ptr,
-        reply_len,
-        control: control_ptr,
-        control_len,
-        error: ptr::null_mut(),
-    }
-}
-
-fn into_kvpairs(pairs: Vec<(&'static str, String)>) -> Vec<KVPair> {
-    #[allow(clippy::expect_used)]
-    pairs
-        .into_iter()
-        .map(|(k, v)| {
-            let k_c = CString::new(k).expect("literal CString");
-            let v_c = CString::new(v).expect("literal CString");
-            KVPair {
-                key: k_c.into_raw(),
-                value: v_c.into_raw(),
-            }
-        })
-        .collect()
-}
-
-/// Create an AuthError result with the given message. Might panic if `message` can't be turned into a C string, but since we only call this with static strings or error messages from Rust code, that should be fine.
-fn auth_error(response: &Response, message: String) -> AuthResultC {
-    // At some point we just have to convert to a C string, and if that fails we can't do much about it, so it's fine to panic with a literal here
-    #[allow(clippy::expect_used)]
-    let c_message = CString::new(message)
-        .unwrap_or_else(|_| CString::new("module error").expect("literal CString"));
-    AuthResultC {
-        code: response.code(),
-        reply: ptr::null_mut(),
-        reply_len: 0,
-        control: ptr::null_mut(),
-        control_len: 0,
-        error: c_message.into_raw(),
-    }
-}
-
-/// Instantiate module state from a config path.
-#[unsafe(no_mangle)]
-pub extern "C" fn rlm_kanidm_instantiate(config_path: *const c_char) -> *mut ModuleHandle {
-    let Ok(path) = cstr_to_string(config_path) else {
-        return ptr::null_mut();
-    };
-    match Module::from_config_path(&path, &ModuleOptions::default()) {
-        Ok(module) => {
-            let handle = ModuleHandle { module };
-            Box::into_raw(Box::new(handle))
-        }
-        Err(_) => ptr::null_mut(),
-    }
-}
-
-/// Free module state.
-///
-/// # Safety
-/// - if you're calling this from C, ensure you only call it once per handle returned from `rlm_kanidm_instantiate` and do not use the handle after calling this.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn rlm_kanidm_detach(handle: *mut ModuleHandle) {
-    if handle.is_null() {
-        return;
-    }
-    unsafe {
-        drop(Box::from_raw(handle));
-    }
-}
-
-/// Authorize a request represented as key/value pairs.
-///
-/// # Safety
-/// - `handle` must be a pointer returned from `rlm_kanidm_instantiate` and not yet freed by `rlm_kanidm_detach`.
-/// - `request_attrs` must point to an array of `KVPair` of length `request_attrs_len`.
-/// - The strings pointed to by `KVPair` must be valid null-terminated C strings.
-#[unsafe(no_mangle)]
-pub unsafe extern "C" fn rlm_kanidm_authorize(
-    handle: *mut ModuleHandle,
-    request_attrs: *const KVPair,
-    request_attrs_len: usize,
-) -> AuthResultC {
-    if handle.is_null() {
-        return auth_error(&Response::Fail, "null module handle".to_string());
-    }
-
-    let attrs = match kvpairs_to_attributes(request_attrs, request_attrs_len) {
-        Ok(v) => v,
-        Err(e) => return e.into(),
-    };
-
-    let module = unsafe { &(*handle).module };
-    auth_result_from_pairs(&module.authorize(&attrs))
-}
-
-/// Free memory allocated in `AuthResultC`.
-#[unsafe(no_mangle)]
-pub extern "C" fn rlm_kanidm_free_auth_result(result: AuthResultC) {
-    free_kv_pairs(result.reply, result.reply_len);
-    free_kv_pairs(result.control, result.control_len);
-    if !result.error.is_null() {
-        unsafe {
-            drop(CString::from_raw(result.error));
-        }
-    }
-}
-
-/// Helper to free an array of KVPair allocated in Rust and returned to C. This should be called for the `reply` and `control` fields of `AuthResultC` after the caller is done using them, to avoid memory leaks.
-fn free_kv_pairs(ptr_pairs: *mut KVPair, len: usize) {
-    if ptr_pairs.is_null() {
-        return;
-    }
-    let slice_ptr = ptr::slice_from_raw_parts_mut(ptr_pairs, len);
-    let boxed = unsafe { Box::from_raw(slice_ptr) };
-    for pair in boxed.iter() {
-        if !pair.key.is_null() {
-            unsafe {
-                drop(CString::from_raw(pair.key.cast_mut()));
-            }
-        }
-        if !pair.value.is_null() {
-            unsafe {
-                drop(CString::from_raw(pair.value.cast_mut()));
-            }
-        }
-    }
-}
-
-fn kvpairs_to_attributes(
-    request_attrs: *const KVPair,
-    request_attrs_len: usize,
-) -> Result<RequestAttributes, ModuleError> {
-    if request_attrs.is_null() && request_attrs_len != 0 {
-        return Err(ModuleError::Other(
-            "request_attrs pointer is null with non-zero length".to_string(),
-        ));
-    }
-    let attrs_slice: &[KVPair] = if request_attrs_len != 0 {
-        unsafe { std::slice::from_raw_parts(request_attrs, request_attrs_len) }
-    } else {
-        &[]
-    };
-    let mut attrs = BTreeMap::<String, String>::new();
-    for pair in attrs_slice {
-        let key = cstr_to_string(pair.key)?;
-        let value = cstr_to_string(pair.value)?;
-        attrs.insert(key, value);
-    }
-    Ok(RequestAttributes { attrs })
 }
 
 #[cfg(test)]
